+++ common/src/main/java/club/icomponent/common/transfer/auth/CommonCredentials.java
@@ -1,4 +1,4 @@
-package club.icomponent.common.transfer.auth.login;
+package club.icomponent.common.transfer.auth;
 
 import club.icomponent.core.spec.Validation;
 import club.icomponent.core.util.StringUtils;
